SendCryptoAddressTextField refactored
