+++ VultisigApp/VultisigApp/Views/Components/TextFields/SendCryptoAddressTextField.swift
@@ -9,10 +9,6 @@ import SwiftUI
 import OSLog
 import UniformTypeIdentifiers
 
-#if os(iOS)
-import CodeScanner
-#endif
-
 struct SendCryptoAddressTextField: View {
     @ObservedObject var tx: SendTransaction
     @ObservedObject var sendCryptoViewModel: SendCryptoViewModel
@@ -28,37 +24,7 @@ struct SendCryptoAddressTextField: View {
 #endif
     
     var body: some View {
-        content
-            .font(.body12Menlo)
-            .foregroundColor(.neutral0)
-            .frame(height: 48)
-            .frame(maxWidth: .infinity, alignment: .leading)
-            .background(Color.blue600)
-            .cornerRadius(10)
-#if os(iOS)
-            .sheet(isPresented: $showScanner) {
-                codeScanner
-            }
-            .sheet(isPresented: $showImagePicker, onDismiss: processImage) {
-                ImagePicker(selectedImage: $selectedImage)
-            }
-#elseif os(macOS)
-            .fileImporter(
-                isPresented: $showImagePicker,
-                allowedContentTypes: [UTType.image],
-                allowsMultipleSelection: false
-            ) { result in
-                do {
-                    let qrCode = try Utils.handleQrCodeFromImage(result: result)
-                    handleImageQrCode(data: qrCode)
-                } catch {
-                    print(error)
-                }
-            }
-            .onDrop(of: [.image], isTargeted: $isUploading) { providers -> Bool in
-                OnDropQRUtils.handleOnDrop(providers: providers, handleImageQrCode: handleImageQrCode)
-            }
-#endif
+        container
     }
     
     var content: some View {
@@ -88,53 +54,6 @@ struct SendCryptoAddressTextField: View {
                 .foregroundColor(.neutral0)
         }
     }
-    
-    var field: some View {
-        HStack(spacing: 0) {
-            TextField(NSLocalizedString("enterAddress", comment: "").capitalized, text: Binding<String>(
-                get: { tx.toAddress },
-                set: { newValue in
-                    tx.toAddress = newValue
-                    DebounceHelper.shared.debounce {
-                        validateAddress(newValue)
-                    }
-                }
-            ))
-            .foregroundColor(.neutral0)
-            .submitLabel(.next)
-            .disableAutocorrection(true)
-            .borderlessTextFieldStyle()
-            .maxLength(Binding<String>(
-                get: { tx.toAddress },
-                set: { newValue in
-                    tx.toAddress = newValue
-                    DebounceHelper.shared.debounce {
-                        validateAddress(newValue)
-                    }
-                }
-            ))
-#if os(iOS)
-            .keyboardType(.default)
-            .textInputAutocapitalization(.never)
-            .textContentType(.oneTimeCode)
-#endif
-            
-            pasteButton
-#if os(iOS)
-            scanButton
-#elseif os(macOS)
-            fileButton
-#endif
-            addressBookButton
-        }
-        .padding(.horizontal, 12)
-    }
-    
-#if os(iOS)
-    var codeScanner: some View {
-        QRCodeScannerView(showScanner: $showScanner, handleScan: handleScan)
-    }
-#endif
     
     var pasteButton: some View {
         Button {
@@ -184,92 +103,9 @@ struct SendCryptoAddressTextField: View {
         }
     }
     
-    private func processImage() {
-        guard let selectedImage = selectedImage else { return }
-        
-#if os(iOS)
-        handleImageQrCode(image: selectedImage)
-#endif
-    }
-    
-    private func validateAddress(_ newValue: String) {
+    func validateAddress(_ newValue: String) {
         sendCryptoViewModel.validateAddress(tx: tx, address: newValue)
     }
-    
-    private func pasteAddress() {
-#if os(iOS)
-        if let clipboardContent = UIPasteboard.general.string {
-            tx.toAddress = clipboardContent
-            
-            DebounceHelper.shared.debounce {
-                validateAddress(clipboardContent)
-            }
-        }
-#elseif os(macOS)
-        let pasteboard = NSPasteboard.general
-        if let clipboardContent = pasteboard.string(forType: .string) {
-            tx.toAddress = clipboardContent
-            
-            DebounceHelper.shared.debounce {
-                validateAddress(clipboardContent)
-            }
-        }
-#endif
-    }
-    
-#if os(iOS)
-    private func handleScan(result: Result<ScanResult, ScanError>) {
-        switch result {
-        case .success(let result):
-            let qrCodeResult = result.string
-            tx.parseCryptoURI(qrCodeResult)
-            validateAddress(tx.toAddress)
-            showScanner = false
-        case .failure(let err):
-            sendCryptoViewModel.logger.error("fail to scan QR code,error:\(err.localizedDescription)")
-        }
-    }
-    
-    private func handleImageQrCode(image: UIImage) {
-        
-        let qrCodeFromImage = Utils.handleQrCodeFromImage(image: image)
-        let (address, amount, message) = Utils.parseCryptoURI(String(data: qrCodeFromImage, encoding: .utf8) ?? .empty)
-        
-        tx.toAddress = address
-        tx.amount = amount
-        tx.memo = message
-        
-        DebounceHelper.shared.debounce {
-            validateAddress(address)
-        }
-        
-        Task{
-            if !amount.isEmpty {
-                await sendCryptoViewModel.convertToFiat(newValue: amount, tx: tx)
-            }
-        }
-    }
-#elseif os(macOS)
-    private func handleImageQrCode(data: Data) {
-        
-        let (address, amount, message) = Utils.parseCryptoURI(String(data: data, encoding: .utf8) ?? .empty)
-        
-        tx.toAddress = address
-        tx.amount = amount
-        tx.memo = message
-        
-        DebounceHelper.shared.debounce {
-            validateAddress(address)
-        }
-        
-        Task{
-            if !amount.isEmpty {
-                await sendCryptoViewModel.convertToFiat(newValue: amount, tx: tx)
-            }
-        }
-    }
-#endif
-    
 }
 
 #Preview {
